limit(n) now takes a Long instead of Int

+++ akka-stream/src/main/scala/akka/stream/StreamLimitReachedException.scala
@@ -1,3 +1,3 @@
 package akka.stream
 
-class StreamLimitReachedException(val n: Int) extends RuntimeException(s"limit of $n reached")
+class StreamLimitReachedException(val n: Long) extends RuntimeException(s"limit of $n reached")
